Refactoring: GridCreatingConditionTriangle::MovePolygonVertexCommand implementation is separated from gridcreatingconditiontriangle.cpp.
