+++ libs/gridcreatingcondition/triangle/gridcreatingconditiontriangle.cpp
@@ -13,6 +13,7 @@
 #include "private/gridcreatingconditiontriangle_finishpolygondefiningcommand.h"
 #include "private/gridcreatingconditiontriangle_finishpolylinedefiningcommand.h"
 #include "private/gridcreatingconditiontriangle_movepolygoncommand.h"
+#include "private/gridcreatingconditiontriangle_movepolygonvertexcommand.h"
 
 #include <guibase/widget/waitdialog.h>
 #include <guicore/base/iricmainwindowinterface.h>
@@ -271,76 +272,6 @@ void GridCreatingConditionTriangle::mouseDoubleClickEvent(QMouseEvent* /*event*/
 		}
 	}
 }
-
-class GridCreatingConditionTriangle::MovePolygonVertexCommand : public QUndoCommand
-{
-public:
-	MovePolygonVertexCommand(bool keyDown, const QPoint& from, const QPoint& to, vtkIdType vertexId, GridCreatingConditionTriangle* pol) :
-		QUndoCommand {GridCreatingConditionTriangle::tr("Move Polygon Vertex")}
-	{
-		m_keyDown = keyDown;
-		m_vertexId = vertexId;
-		double dx = from.x();
-		double dy = from.y();
-		pol->graphicsView()->viewportToWorld(dx, dy);
-		QVector2D fromVec(dx, dy);
-		dx = to.x();
-		dy = to.y();
-		pol->graphicsView()->viewportToWorld(dx, dy);
-		QVector2D toVec(dx, dy);
-		m_offset = toVec - fromVec;
-		m_polygon = pol;
-		m_targetPolygon = m_polygon->m_selectedPolygon;
-	}
-	void redo() {
-		vtkPolygon* pol = m_targetPolygon->getVtkPolygon();
-		vtkPoints* points = pol->GetPoints();
-		double p[3];
-		points->GetPoint(m_vertexId, p);
-		p[0] += m_offset.x();
-		p[1] += m_offset.y();
-		points->SetPoint(m_vertexId, p);
-
-		points->Modified();
-		pol->Modified();
-		m_targetPolygon->updateShapeData();
-		m_polygon->renderGraphicsView();
-	}
-	void undo() {
-		vtkPolygon* pol = m_targetPolygon->getVtkPolygon();
-		vtkPoints* points = pol->GetPoints();
-		double p[3];
-		points->GetPoint(m_vertexId, p);
-		p[0] -= m_offset.x();
-		p[1] -= m_offset.y();
-		points->SetPoint(m_vertexId, p);
-
-		points->Modified();
-		pol->Modified();
-		m_targetPolygon->updateShapeData();
-		m_polygon->renderGraphicsView();
-	}
-	int id() const {
-		return iRIC::generateCommandId("MovePolygonVertex");
-	}
-	bool mergeWith(const QUndoCommand* other) {
-		const MovePolygonVertexCommand* comm = dynamic_cast<const MovePolygonVertexCommand*>(other);
-		if (comm == nullptr) {return false;}
-		if (comm->m_keyDown) {return false;}
-		if (comm->m_polygon != m_polygon) {return false;}
-		if (comm->m_targetPolygon != m_targetPolygon) {return false;}
-		if (comm->m_vertexId != m_vertexId) {return false;}
-		m_offset += comm->m_offset;
-		return true;
-	}
-
-private:
-	bool m_keyDown;
-	vtkIdType m_vertexId;
-	GridCreatingConditionTriangle* m_polygon;
-	GridCreatingConditionTriangleAbstractPolygon* m_targetPolygon;
-	QVector2D m_offset;
-};
 
 class GridCreatingConditionTriangle::AddPolygonVertexCommand : public QUndoCommand
 {
