+++ libs/gridcreatingcondition/triangle/triangle.pro
@@ -110,7 +110,8 @@ HEADERS += gcc_triangle_global.h \
            private/gridcreatingconditiontriangle_definepolygonnewpointcommand.h \
            private/gridcreatingconditiontriangle_finishpolygondefiningcommand.h \
            private/gridcreatingconditiontriangle_finishpolylinedefiningcommand.h \
-           private/gridcreatingconditiontriangle_movepolygoncommand.h
+           private/gridcreatingconditiontriangle_movepolygoncommand.h \
+           private/gridcreatingconditiontriangle_movepolygonvertexcommand.h
 FORMS += gridcreatingconditiontrianglepolygoncoordinateseditdialog.ui \
          gridcreatingconditiontrianglepolylinecoordinateseditdialog.ui \
          gridcreatingconditiontriangleremeshdialog.ui \
@@ -131,7 +132,8 @@ SOURCES += gridcreatingconditioncreatortriangle.cpp \
            private/gridcreatingconditiontriangle_definepolygonnewpointcommand.cpp \
            private/gridcreatingconditiontriangle_finishpolygondefiningcommand.cpp \
            private/gridcreatingconditiontriangle_finishpolylinedefiningcommand.cpp \
-           private/gridcreatingconditiontriangle_movepolygoncommand.cpp
+           private/gridcreatingconditiontriangle_movepolygoncommand.cpp \
+           private/gridcreatingconditiontriangle_movepolygonvertexcommand.cpp
 TRANSLATIONS += languages/iricGccTriangle_es_ES.ts \
                 languages/iricGccTriangle_fr_FR.ts \
                 languages/iricGccTriangle_id_ID.ts \
